A LOT OF THINGS
Added bullet points for the achievements
and the ability to reset acheivement values in game

Completely reworked the plane throwing because it was buggy: it would try and correct its course to the hoop which made it difficult to predict the oscillating hoops movement.
The planes also used velocity so they didn't always go in the direction plane was facing.

I also changed the UI a little bit. Moved the buttons closer/down and made them smaller(needed room for achievments).

Added score and achievements: works the same way as balloon game. (level is score)