Added auto changing difficulty
The difficulty now changes with the players progress. however it might need tweaking because, unlike the balloon game where it only controls speed the difficulty is used for oscilation, range, and Z positioning and could get out of hand.

I also remove the kinect. The code is still there but until we get the server working I reimplemented the touch to make it playable